Code rearrangement to make the code a module and not a script. #5 #9 #10
